overlay: Enable deskDockEnablesAccelerometer
Control whether being in the desk dock should enable accelerometer
based screen orientation. This defaults to true because it is
common for desk docks to be sold in a variety of form factors
with different orientations. Since we cannot always tell these docks
apart and the docks cannot report their true orientation on their own,
we rely on gravity to determine the effective orientation.
Default: true (as in stock firmware)